% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessdiff.R
\name{sessdiff}
\alias{sessdiff}
\title{use sessioninfo package to identify differences between session configurations}
\usage{
sessdiff(x, y, ...)
}
\arguments{
\item{x}{instance of session_info from sessioninfo package}

\item{y}{instance of session_info from sessioninfo package}

\item{\dots}{not used.  Could be passed to `session_info` but then there
would need to be code to extract information on, e.g., external components.
This enhancement should be done but probably not in this function.}
}
\description{
use sessioninfo package to identify differences between session configurations
}
\examples{
if (requireNamespace("sessioninfo")) {
  si = sessioninfo::session_info()
  print(sessdiff( si, si ))
  if (requireNamespace("parody")) { # unlikely to be loaded in any session
    si2 = sessioninfo::session_info()
    sessdiff(si, si2)
    }
  }
}
